@@ -42,36 +42,35 @@
   ],
   "homepage": "https://github.com/speech-rule-engine/sre-tests",
   "dependencies": {
-    "@google-cloud/storage": "6.10.1",
+    "@google-cloud/storage": "7.3.2",
     "@xmldom/xmldom": "^0.9.0-beta.8",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-prettier": "^9.0.0",
     "esm": "3.2.25",
-    "firebase-admin": "^11.8.0",
-    "firebase-tools": "^12.2.1",
-    "jest": "^29.6.2",
-    "mathjax-full": "3.2.2",
-    "npm-check-updates": "16.10.12",
-    "sre-visualiser": "^0.5.0",
+    "firebase-admin": "^11.11.0",
+    "firebase-tools": "^12.7.0",
+    "jest": "^29.7.0",
+    "mathjax-full": "4.0.0-beta.4",
     "wicked-good-xpath": "1.3.0"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.22.4",
-    "@types/jest": "29.5.1",
-    "@types/node": "20.2.5",
-    "@typescript-eslint/eslint-plugin": "5.59.8",
-    "@typescript-eslint/parser": "5.59.8",
-    "babel-jest": "29.5.0",
+    "@babel/preset-env": "7.23.2",
+    "@types/jest": "29.5.6",
+    "@types/node": "20.8.9",
+    "@typescript-eslint/eslint-plugin": "6.9.0",
+    "@typescript-eslint/parser": "6.9.0",
+    "babel-jest": "29.7.0",
     "babel-plugin-transform-class-properties": "6.24.1",
-    "eslint": "8.41.0",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jsdoc": "45.0.0",
-    "prettier": "^2.8.8",
+    "eslint": "8.52.0",
+    "eslint-plugin-import": "2.29.0",
+    "eslint-plugin-jsdoc": "46.8.2",
+    "prettier": "^3.0.3",
+    "rimraf": "^5.0.5",
     "terser-webpack-plugin": "5.3.9",
-    "ts-loader": "9.4.3",
-    "typescript": "5.0.4",
+    "ts-loader": "9.5.0",
+    "typescript": "5.2.2",
     "typescript-tools": "0.3.1",
-    "webpack": "5.84.1",
-    "webpack-cli": "5.1.1"
+    "webpack": "5.89.0",
+    "webpack-cli": "5.1.4"
   },
   "files": [
     "LICENSE",
